dtr3

+++ DQNTRADINGBGOT/Agent/TD3/buffer.py
@@ -1,14 +1,13 @@
 import numpy as np
 
-
 class ReplayBuffer():
     def __init__(self, max_size, input_shape, n_actions):
         self.mem_size = max_size
         self.mem_cntr = 0
-        self.state_memory = np.zeros((self.mem_size, *input_shape),  dtype=np.float32)
-        self.new_state_memory = np.zeros((self.mem_size, *input_shape) ,dtype=np.float32)
+        self.state_memory = np.zeros((self.mem_size, *input_shape))
+        self.new_state_memory = np.zeros((self.mem_size, *input_shape))
         self.action_memory = np.zeros((self.mem_size, n_actions))
-        self.reward_memory = np.zeros(self.mem_size,  dtype=np.float32)
+        self.reward_memory = np.zeros(self.mem_size)
         self.terminal_memory = np.zeros(self.mem_size, dtype=np.bool)
 
     def store_transition(self, state, action, reward, state_, done):
@@ -33,3 +32,5 @@ class ReplayBuffer():
         dones = self.terminal_memory[batch]
 
         return states, actions, rewards, states_, dones
+
+
